+++ demo/app/symbol/symbol-demo.component.ts
@@ -1,0 +1,16 @@
+import { Component, OnInit } from '@angular/core';
+import { SymbolService } from '../../../packages/ui/src/symbol/symbol.service';
+
+@Component({
+    selector: 'ec-symbol-demo',
+    templateUrl: 'symbol-demo.component.html'
+})
+
+export class SymbolDemoComponent implements OnInit {
+    constructor(public symbol: SymbolService) {
+
+        console.log('symbol', this.symbol.resolve('entry_saved'));
+    }
+
+    ngOnInit() { }
+}