+++ packages/ui/src/symbol/de.ts
@@ -1,0 +1,9 @@
+import { Symbol } from './symbol.interface';
+
+export const locales: Symbol[] = [{
+    name: 'entry_saved',
+    content: 'Der Eintrag wurde gespeichert'
+}, {
+    name: 'save_error',
+    content: 'Fehler beim Speichern'
+}];