+++ packages/ui/src/symbol/symbol.module.ts
@@ -1,0 +1,11 @@
+import { NgModule } from '@angular/core';
+import { CommonModule } from '@angular/common';
+import { SymbolService } from './symbol.service';
+
+@NgModule({
+    imports: [CommonModule],
+    declarations: [],
+    exports: [],
+    providers: [SymbolService],
+})
+export class SymbolModule { }